refactor: renamed assets
